+++ server/domain/src/form.rs
@@ -1,1 +1,5 @@
+pub mod answer;
+pub mod comment;
+pub mod message;
 pub mod models;
+pub mod question;
