+++ server/domain/src/form/answer/models.rs
@@ -1,0 +1,37 @@
+use chrono::{DateTime, Utc};
+#[cfg(test)]
+use proptest_derive::Arbitrary;
+use serde::{Deserialize, Serialize};
+
+use crate::{
+    form::{models::FormId, question::models::QuestionId},
+    user::models::User,
+};
+
+pub type AnswerId = types::IntegerId<FormAnswer>;
+
+#[derive(Serialize, Deserialize, PartialEq, Debug)]
+pub struct FormAnswer {
+    pub id: AnswerId,
+    pub user: User,
+    pub timestamp: DateTime<Utc>,
+    pub form_id: FormId,
+    pub title: Option<String>,
+}
+
+#[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
+pub struct FormAnswerContent {
+    pub answer_id: AnswerId,
+    pub question_id: QuestionId,
+    pub answer: String,
+}
+
+pub type AnswerLabelId = types::IntegerId<AnswerLabel>;
+
+#[cfg_attr(test, derive(Arbitrary))]
+#[derive(Serialize, Deserialize, Debug, PartialEq)]
+pub struct AnswerLabel {
+    pub id: AnswerLabelId,
+    pub answer_id: AnswerId,
+    pub name: String,
+}
