+++ server/domain/src/form/message.rs
@@ -1,0 +1,1 @@
+pub mod models;
