+++ server/domain/src/repository/form_repository.rs
@@ -3,10 +3,15 @@ use errors::Error;
 use mockall::automock;
 
 use crate::{
-    form::models::{
-        AnswerId, AnswerLabel, Comment, CommentId, DefaultAnswerTitle, Form, FormAnswer,
-        FormAnswerContent, FormDescription, FormId, FormTitle, Label, LabelId, Message, MessageId,
-        Question, ResponsePeriod, Visibility, WebhookUrl,
+    form::{
+        answer::models::{AnswerId, AnswerLabel, FormAnswer, FormAnswerContent},
+        comment::models::{Comment, CommentId},
+        message::models::{Message, MessageId},
+        models::{
+            DefaultAnswerTitle, Form, FormDescription, FormId, FormTitle, Label, LabelId,
+            ResponsePeriod, Visibility, WebhookUrl,
+        },
+        question::models::Question,
     },
     types::authorization_guard::{AuthorizationGuard, Create, Delete, Read, Update},
     user::models::User,
