+++ server/infra/resource/src/database/search.rs
@@ -1,6 +1,9 @@
 use async_trait::async_trait;
 use domain::{
-    form::models::{Form, FormAnswerContent, Label},
+    form::{
+        answer::models::FormAnswerContent,
+        models::{Form, Label},
+    },
     search::models::Comment,
     user::models::User,
 };
