+++ server/usecase/src/form.rs
@@ -1,9 +1,14 @@
 use chrono::Utc;
 use domain::{
-    form::models::{
-        AnswerId, Comment, CommentId, DefaultAnswerTitle, Form, FormAnswerContent, FormDescription,
-        FormId, FormTitle, Label, LabelId, Message, MessageId, Question, ResponsePeriod,
-        Visibility, Visibility::PUBLIC, WebhookUrl,
+    form::{
+        answer::models::{AnswerId, FormAnswerContent},
+        comment::models::{Comment, CommentId},
+        message::models::{Message, MessageId},
+        models::{
+            DefaultAnswerTitle, Form, FormDescription, FormId, FormTitle, Label, LabelId,
+            ResponsePeriod, Visibility, Visibility::PUBLIC, WebhookUrl,
+        },
+        question::models::Question,
     },
     notification::models::{Notification, NotificationSource},
     repository::{
